backend/requirements.txt
Django>=3.0,<4.0
python-decouple==3.4
psycopg2-binary==2.8
django-cors-headers==3.7.0
django-heroku==0.3.1
dj-database-url==0.5.0
whitenoise==5.2
gunicorn==20.0
dj-static==0.0.6
robotframework
robotframework-seleniumlibrary
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt

# lib for chat and celery
celery
redis
django-celery-results
channels
channels_redis
